Update Navbar.jsx
The navigation bar (navbar) is intended to be fixed at the top of the viewport when scrolling, but currently, there is a visible gap between the top of the navbar and the top of the viewport. This gap allows the content behind the navbar to be visible, creating an unintended space and disrupting the layout and user experience. So I slove that problem.